impl: VM作成のワークフローを追加
