+++ kernel/src/model/mod.rs
@@ -1,0 +1,3 @@
+pub mod template;
+pub mod user;
+pub mod virtual_machine;
