+++ kernel/src/model/user/mod.rs
@@ -1,0 +1,10 @@
+use chrono::{DateTime, Utc};
+use uuid::Uuid;
+
+pub struct User {
+    pub id: Uuid,
+    pub name: String,
+    pub email: String,
+    pub created_at: DateTime<Utc>,
+    pub update_at: DateTime<Utc>,
+}
