+++ kernel/src/model/virtual_machine/event.rs
@@ -1,0 +1,7 @@
+use uuid::Uuid;
+
+pub struct CreateVirtualMachine {
+    pub user_id: Uuid,
+    pub template_id: Uuid,
+    pub name: String,
+}
