+++ kernel/src/repository/mod.rs
@@ -1,0 +1,2 @@
+pub mod proxmox_api;
+pub mod virtual_machine;
