+++ kernel/src/repository/proxmox_api.rs
@@ -1,0 +1,5 @@
+use crate::model::virtual_machine::VirtualMachine;
+
+pub trait ProxmoxApiRepository {
+    fn create_vm(&self, virtual_machine: VirtualMachine) -> ();
+}
