+++ kernel/src/repository/virtual_machine.rs
@@ -1,0 +1,4 @@
+use crate::model::virtual_machine::{event::CreateVirtualMachine, VirtualMachine};
+pub trait VirtualMachineRepository {
+    fn insert(&self, event: CreateVirtualMachine) -> VirtualMachine;
+}
